test: move the environment probes into lib/checks.py
The runner owns discovery, containment and grading; what the build was
configured with and what the kernel offers is a different question, and
run-tests.py had grown four answers to it.
